fixed neighbourhood connectivity changed protein.deg to int stub for loading GO information
git-svn-id: https://ontologizer.svn.sourceforge.net/svnroot/ontologizer@665 07a4415e-894b-0410-a0d6-e7c924882cef

+++ branches/ontologizer/graph-measures/src/sonumina/math/graph/DirectedGraph.java
@@ -1363,7 +1363,7 @@ public class DirectedGraph<VertexType> extends AbstractGraph<VertexType> impleme
 			for(VertexType v: sample)
 				measure += vertexMeasure.getMeasure(v);
 
-			sampledAvgMeasure[i] = (double) measure/sampleSize;
+			sampledAvgMeasure[i] = (double) measure/ (double) sampleSize;
 
 		}
 		Arrays.sort(sampledAvgMeasure);
